requirements.txt
selenium
langchain
chromadb[bm25]
nltk
pytest
langchain-community
langchain_ollama
langchain-chroma
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.2
huggingface_hub==0.16.4
rank_bm25
together
deep-translator
python-dotenv
streamlit
fastapi
uvicorn